Moved package into pkg and renamed importer -> interpreter package.
