+++ .gitignore
@@ -1,3 +1,3 @@
 /.idea/
 /bin/
-/interpreter/testdata/dest
+/pkg/interpreter/testdata/dest
