+++ pkg/interpreter/foreach.go
@@ -1,4 +1,4 @@
-package importer
+package interpreter
 
 import (
 	"errors"
