+++ pkg/interpreter/functions.go
@@ -1,4 +1,4 @@
-package importer
+package interpreter
 
 import (
 	"bytes"
