+++ pkg/interpreter/ignore.go
@@ -1,4 +1,4 @@
-package importer
+package interpreter
 
 const (
 	ignoreStart = "start"
