+++ pkg/interpreter/interpreter_test.go
@@ -1,4 +1,4 @@
-package importer
+package interpreter
 
 import (
 	"archive/zip"
