+++ pkg/interpreter/queue.go
@@ -1,4 +1,4 @@
-package importer
+package interpreter
 
 type queue struct {
 	v []foreachBuffer
