+++ pkg/interpreter/testdata/fastplate.var
