+++ pkg/interpreter/testdata/gold/dirMode.sum
