+++ pkg/interpreter/testdata/gold/fileMode.sum
