+++ pkg/interpreter/testdata/src/partials/functions.yaml
