+++ pkg/interpreter/testdata/src/partials/nested.yaml
