+++ pkg/interpreter/testdata/src/partials/nested2.yaml
