+++ pkg/interpreter/testdata/src/partials/varscopes.yaml
